+++ src/main/java/com/wiley/pdh/pdhproductionsupporter/entity/ProductDatesXref.java
@@ -26,19 +26,19 @@ public class ProductDatesXref {
     @Column(name = "PKEY_SRC_OBJECT")
     private String pkeySrcObject;
     @Column(name = "C_DH_ID")
-    private int cDhId;
+    private BigInteger cDhId;
     @Column(name = "ROWID_SYSTEM")
     private String rowIdSystem;
     @Column(name = "SRC_LUD")
     private java.sql.Date srcLud;
     @Column(name = "PUT_UPDATE_MERGE_IND")
-    private int putUpdateMergeInd;
+    private Integer putUpdateMergeInd;
     @Column(name = "INTERACTION_ID")
-    private int interactionId;
+    private BigInteger interactionId;
     @Column(name = "HUB_STATE_IND")
-    private int hubStateInd;
+    private Integer hubStateInd;
     @Column(name = "PROMOTE_IND")
-    private int promoteInd;
+    private Integer promoteInd;
     @Column(name = "NULL_INDICATOR_BITMAP")
     private String nullIndicatorBitMap;
     @Column(name = "CREATOR")
@@ -56,9 +56,9 @@ public class ProductDatesXref {
     @Column(name = "DELETED_DATE")
     private Timestamp deletedDate;
     @Column(name = "DH_PRODREG_ID")
-    private int dhProdRegId;
+    private BigInteger dhProdRegId;
     @Column(name = "DH_DATE_TYPE_ID")
-    private int dhDateTypeId;
+    private BigInteger dhDateTypeId;
     @Column(name = "DATE_VALUE")
     private java.sql.Date dateValue;
     @Column(name = "START_DATE")
@@ -68,7 +68,7 @@ public class ProductDatesXref {
     @Column(name = "PRIMARY_IND")
     private String primaryInd;
     @Column(name = "DH_PRODTYPE_ID")
-    private int dhProdTypeId;
+    private BigInteger dhProdTypeId;
     @Column(name = "PRODREG_SRC_KEY")
     private String prodRegSrcKey;
     @Column(name = "EXCEPTION_FLAG")
